+++ ff/src/biginteger/mod.rs
@@ -85,7 +85,7 @@ macro_rules! const_modulo {
             remainder = remainder.const_mul2();
             remainder.0[0] |= $a.get_bit(i as usize) as u64;
             if remainder.const_geq($divisor) {
-                let (r, borrow) = remainder.const_sub_noborrow($divisor);
+                let (r, borrow) = remainder.const_sub_with_borrow($divisor);
                 remainder = r;
                 assert!(!borrow);
             }
@@ -183,7 +183,7 @@ impl<const N: usize> BigInt<N> {
 
     #[inline]
     #[ark_ff_asm::unroll_for_loops]
-    pub(crate) const fn const_sub_noborrow(mut self, other: &Self) -> (Self, bool) {
+    pub(crate) const fn const_sub_with_borrow(mut self, other: &Self) -> (Self, bool) {
         let mut borrow = 0;
 
         const_for!((i in 0..N) {
@@ -235,7 +235,7 @@ impl<const N: usize> BigInteger for BigInt<N> {
 
     #[inline]
     #[ark_ff_asm::unroll_for_loops]
-    fn add_nocarry(&mut self, other: &Self) -> bool {
+    fn add_with_carry(&mut self, other: &Self) -> bool {
         let mut carry = 0;
 
         for i in 0..N {
@@ -257,7 +257,7 @@ impl<const N: usize> BigInteger for BigInt<N> {
 
     #[inline]
     #[ark_ff_asm::unroll_for_loops]
-    fn sub_noborrow(&mut self, other: &Self) -> bool {
+    fn sub_with_borrow(&mut self, other: &Self) -> bool {
         let mut borrow = 0;
 
         for i in 0..N {
@@ -701,42 +701,47 @@ pub trait BigInteger:
     /// Number of 64-bit limbs representing `Self`.
     const NUM_LIMBS: usize;
 
-    /// Add another representation to this one, returning the carry bit.
+    /// Add another [`BigInteger`] to `self`. This method stores the result in `self`,
+    /// and returns a carry bit.
     /// # Example
     ///
     /// ```
     /// use ark_ff::{biginteger::BigInteger64 as B, BigInteger as _};
     ///
     /// // Basic
-    /// let (mut one, mut two_add) = (B::from(1u64), B::from(2u64));
-    /// two_add.add_nocarry(&one);
-    /// assert_eq!(two_add, B::from(3u64));
+    /// let (mut one, mut x) = (B::from(1u64), B::from(2u64));
+    /// let carry = x.add_with_carry(&one);
+    /// assert_eq!(x, B::from(3u64));
+    /// assert_eq!(carry, false);
     ///
     /// // Edge-Case
-    /// let mut max_one = B::from(u64::MAX);
-    /// let carry = max_one.add_nocarry(&one);
-    /// assert_eq!(max_one, B::from(0u64));
+    /// let mut x = B::from(u64::MAX);
+    /// let carry = x.add_with_carry(&one);
+    /// assert_eq!(x, B::from(0u64));
     /// assert_eq!(carry, true)
     /// ```
-    fn add_nocarry(&mut self, other: &Self) -> bool;
+    fn add_with_carry(&mut self, other: &Self) -> bool;
 
-    /// Subtract another representation from this one, returning the borrow bit.
+    /// Subtract another [`BigInteger`] from this one. This method stores the result in
+    /// `self`, and returns a borrow.
+    /// the borrow bit.
     /// # Example
     ///
     /// ```
     /// use ark_ff::{biginteger::BigInteger64 as B, BigInteger as _};
     ///
     /// // Basic
-    /// let (mut one, mut two, mut three_sub) = (B::from(1u64), B::from(2u64), B::from(3u64));
-    /// three_sub.sub_noborrow(&two);
-    /// assert_eq!(three_sub, one);
+    /// let (mut one_sub, two, mut three_sub) = (B::from(1u64), B::from(2u64), B::from(3u64));
+    /// let borrow = three_sub.sub_with_borrow(&two);
+    /// assert_eq!(three_sub, one_sub);
+    /// assert_eq!(borrow, false);
     ///
     /// // Edge-Case
-    /// let borrow = one.sub_noborrow(&two);
+    /// let borrow = one_sub.sub_with_borrow(&two);
+    /// assert_eq!(one_sub, B::from(u64::MAX));
     /// assert_eq!(borrow, true);
-    /// assert_eq!(one, B::from(u64::MAX));
     /// ```
-    fn sub_noborrow(&mut self, other: &Self) -> bool;
+    fn sub_with_borrow(&mut self, other: &Self) -> bool;
 
     /// Performs a leftwise bitshift of this number, effectively multiplying
     /// it by 2. Overflow is ignored.
@@ -763,7 +768,8 @@ pub trait BigInteger:
     /// ```
     fn mul2(&mut self);
 
-    /// Performs a leftwise bitshift of this number by some amount.
+    /// Performs a leftwise bitshift of this number by n bits, effectively multiplying
+    /// it by 2^n. Overflow is ignored.
     /// # Example
     ///
     /// ```
@@ -996,9 +1002,9 @@ pub trait BigInteger:
                 if e.is_odd() {
                     z = signed_mod_reduction(e.as_ref()[0], 1 << w);
                     if z >= 0 {
-                        e.sub_noborrow(&Self::from(z as u64));
+                        e.sub_with_borrow(&Self::from(z as u64));
                     } else {
-                        e.add_nocarry(&Self::from((-z) as u64));
+                        e.add_with_carry(&Self::from((-z) as u64));
                     }
                 } else {
                     z = 0;